Create Blog - frontend
We have fixed the text editor and added all the validations for the images, and created a maximum of 20 blocks - 10 block of images .

We have new validations for the  other inputs.

We send the editor images using base64 encode to backend, and the cover image as normal file.

We have added isDraft flag for drafts.

Bugs to be fixed : rich text editor for bigger images - loading in a loop -- has to be fixed even if it return an error, it doesn t stop .